requirements.txt
langchain==0.3.1
langchain-huggingface==0.1.0
langchain-community==0.3.1
kaggle==1.6.17
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
gradio==4.44.0
python-dotenv==1.0.1
requests==2.32.3
transformers>=4.40.0
torch>=2.1
accelerate